+++ firebase_rules/lib/src/database/model/auth.dart
@@ -28,6 +28,9 @@ enum RulesProvider {
 abstract class RulesToken {
   RulesToken._();
 
+  /// Access to custom claims
+  dynamic operator [](String key);
+
   /// The email address associated with the account, if present
   RulesString get email;