+++ firebase_rules/CHANGELOG.md
@@ -1,5 +1,7 @@
 ## NEXT
 - Renames `FirestorePath` to `FirestoreRoot` and `StoragePath` to `StorageRoot`
+- `RuleDataSnapshot.val()` is now `RuleDataSnapshot.val<T>()` to allow for type-safety
+- Adds a `[]` operator to `RulesToken` to allow access to custom claims
 
 ## 0.0.1+2
 - Readme fixes